requirements.txt
Django==4.2.2
djangorestframework==3.14.0
GitPython==3.1.31
python-dotenv==1.0.0
requests==2.31.0
requests-toolbelt==1.0.0
responses==0.23.1
sqlparse==0.4.4
userpath==1.8.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
webdriver-manager==3.8.6
